+++ frontend/project/src/admin/Sales/AnnualSales.js
@@ -1,23 +1,27 @@
 import React from 'react';
-import '../css/SalesPage.css'; 
+import '../Styles/SalesPage.css';
 
-const MonthlySalesModal = ({  data, onClose }) => {
-  if (!data) return null;    
-  
-  const { product, sales } = data;
+const AnnualSales = ({ data, onClose }) => {
+  if (!data) return null;
+
+  console.log(data);
+
+  const { product } = data;
   const totalQuantitySold = data.total_sold;
   const totalGrossIncome = data.total_sales;
 
+  console.log(totalQuantitySold, totalGrossIncome);
+
   return (
     <div className="modal-overlay">
-      <div className="modal-content fade-in"> {/* Apply fade-in class */}
+      <div className="modal-content fade-in">   
         <button className="close-button" onClick={onClose}>×</button>
         <div className='modal-header'>
-          <h2>{product.product_name} Monthly Sales</h2>
+          <h2>{product.product_name} Annual Sales</h2>
         </div>
         <div className="modal-card">
           <div className="totals">
-            <p>Total Quantity Sold: {totalQuantitySold}</p>
+            <p>Total Quantity Sold: ${totalQuantitySold}</p>
             <p>Total Gross Income: ${totalGrossIncome.toFixed(2)}</p>
           </div>
         </div>
@@ -26,4 +30,4 @@ const MonthlySalesModal = ({  data, onClose }) => {
   );
 };
 
-export default MonthlySalesModal;
+export default AnnualSales;
