+++ frontend/project/src/admin/Sales/MonthlySales.js
@@ -1,20 +1,19 @@
 import React from 'react';
-import '../css/SalesPage.css';
+import '../Styles/SalesPage.css'; 
 
-const WeeklySalesModal = ({ data, onClose }) => {
+const MonthlySalesModal = ({  data, onClose }) => {
   if (!data) return null;    
   
   const { product } = data;
   const totalQuantitySold = data.total_sold;
   const totalGrossIncome = data.total_sales;
-  
 
   return (
     <div className="modal-overlay">
       <div className="modal-content fade-in"> {/* Apply fade-in class */}
         <button className="close-button" onClick={onClose}>×</button>
         <div className='modal-header'>
-          <h2>{product.product_name} Weekly Sales</h2>
+          <h2>{product.product_name} Monthly Sales</h2>
         </div>
         <div className="modal-card">
           <div className="totals">
@@ -27,4 +26,4 @@ const WeeklySalesModal = ({ data, onClose }) => {
   );
 };
 
-export default WeeklySalesModal;
+export default MonthlySalesModal;
