+++ frontend/project/src/admin/Sales/WeeklySales.js
@@ -1,27 +1,24 @@
 import React from 'react';
-import '../css/SalesPage.css';
+import '../Styles/SalesPage.css';
 
-const AnnualSales = ({ data, onClose }) => {
-  if (!data) return null;
-
-  console.log(data);
-
-  const { product, sales } = data;
+const WeeklySalesModal = ({ data, onClose }) => {
+  if (!data) return null;    
+  
+  const { product } = data;
   const totalQuantitySold = data.total_sold;
   const totalGrossIncome = data.total_sales;
-
-  console.log(totalQuantitySold, totalGrossIncome);
+  
 
   return (
     <div className="modal-overlay">
-      <div className="modal-content fade-in">   
+      <div className="modal-content fade-in"> {/* Apply fade-in class */}
         <button className="close-button" onClick={onClose}>×</button>
         <div className='modal-header'>
-          <h2>{product.product_name} Annual Sales</h2>
+          <h2>{product.product_name} Weekly Sales</h2>
         </div>
         <div className="modal-card">
           <div className="totals">
-            <p>Total Quantity Sold: ${totalQuantitySold}</p>
+            <p>Total Quantity Sold: {totalQuantitySold}</p>
             <p>Total Gross Income: ${totalGrossIncome.toFixed(2)}</p>
           </div>
         </div>
@@ -30,4 +27,4 @@ const AnnualSales = ({ data, onClose }) => {
   );
 };
 
-export default AnnualSales;
+export default WeeklySalesModal;
