+++ frontend/project/src/user/Styles/user_products.css
@@ -2,7 +2,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Vollkorn:ital,wght@0,400..900;1,400..900&display=swap');
 
 .user-products-page {
-    background-image: url('../images/user_bg.png');
+    background-image: url('../../images/user_bg.png');
     background-size: cover;
     background-repeat: no-repeat;
     position: absolute;
@@ -190,8 +190,3 @@
 .notification-message {
     flex: 1;
 }
-
-
-
-
-
