+++ frontend/project/src/user/Styles/userlanding.css
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap');
 
 .user-page {
-    background-image: url('../images/userlp_bg.png');
+    background-image: url('../../images/userlp_bg.png');
     background-size: cover;
     background-repeat: no-repeat;
     position: absolute;
